+++ Staging/output_sql/sa_esms.sql
@@ -2,5 +2,5 @@ CREATE TABLE sa_esms (
 	column1 INTEGER,
 	annee VARCHAR,
 	finess VARCHAR,
-	total REAL
+	total INTEGER
 )
