+++ Staging/output_sql/sa_pmsi.sql
@@ -2,9 +2,9 @@ CREATE TABLE sa_pmsi (
 	column1 INTEGER,
 	mois VARCHAR,
 	finess VARCHAR,
-	somme_de_mco_hors_seances REAL,
-	somme_de_mco_seances REAL,
-	somme_de_ssr REAL,
-	somme_de_had REAL,
-	somme_de_psy REAL
+	somme_de_mco_hors_seances INTEGER,
+	somme_de_mco_seances INTEGER,
+	somme_de_ssr INTEGER,
+	somme_de_had INTEGER,
+	somme_de_psy INTEGER
 )
